requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
anyio==4.10.0
asgiref==3.9.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
backoff==2.2.1
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
chapa-cli==0.0.4
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cron_descriptor==2.0.5
cryptography==45.0.7
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.4
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-crontab==0.7.1
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-graphql-jwt==0.4.0
django-ipware==7.0.1
django-ratelimit==4.1.0
django-redis==6.0.0
django-seed==0.3.1
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
drf-yasg==1.21.10
Faker==37.4.2
Flask==3.1.1
frozenlist==1.7.0
future==1.0.0
geoip2==5.1.0
git-filter-repo==2.47.0
gql==4.0.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
maxminddb==2.8.2
mdurl==0.1.2
multidict==6.6.4
packaging==25.0
parameterized==0.9.0
pillow==11.3.0
pluggy==1.6.0
promise==2.3
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.10
psycopg2>=2.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pyngrok==7.3.0
pyOpenSSL==25.1.0
pytest==8.4.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-ipware==3.0.0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
serializers==0.2.4
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
text-unidecode==1.3
toposort==1.10
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.9.0
yarl==1.20.1
zope.interface==7.2
flake8
